requirements.txt
Django==1.7.6
djangorestframework==3.1.0